+++ src/db/migrations/0040_old_freak.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "thread_participants" ADD COLUMN "current_activity" text;